+++ CHANGELOG.md
@@ -6,5 +6,6 @@ Changelog
 
 - [PADA-142](https://givisiez.atlassian.net/browse/PADA-142): Layout da tela de vacinas
 - [PADA-199](https://givisiez.atlassian.net/browse/PADA-199): Integração com Back-End da tela de vacinas
+- [PADA-201](https://givisiez.atlassian.net/browse/PADA-201): Componentizar o menu lateral da telas de paciente (perfil, fases de tratamento e vacinas)
 
 ## Fix